{{cookiecutter.project_slug}}/requirements/dev.txt

# development mode
# paths relative to location of setup.py
-e ".[test]"

-e {{ cookiecutter.simcore_install_root }}/packages/s3wrapper/
-e {{ cookiecutter.simcore_install_root }}/packages/simcore-sdk/
-e {{ cookiecutter.simcore_install_root }}/packages/director-sdk/python

aiohttp-devtools==0.10.1
autopep8>=1.3.5